+++ CHANGELOG.md
@@ -4,9 +4,8 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
-## [Unreleased]
+## [0.7.0] - 2017-12-07
 
-### Changed
 - Configuration format for whitelisting users has been changed to make it easier to
   add further global filters
 - enable and disable filters via the config